Basically:
- Allow immediate child (>) combinator to get quickly rejected by the ancestor filter.
- Bucket div#foo and div.foo as ID/class respectively, rather than as tag(div)
- Bucket :is(.foo) and :where(.foo) as class rules rather than as generic rules

Average full style update times on my machine:
- https://vercel.com 470ms -> 240ms
- https://tailwindcss.com 1400ms -> 350ms